[SPARK-52705][SQL] Refactor deterministic check for grouping expressions
### What changes were proposed in this pull request?

Move check for non-deterministic expressions in grouping expressions from `ExprUtils` to `CheckAnalysis`.
### Why are the changes needed?
This is necessary in order to be able to utilize `PullOutNonDeterminstic` rule as a post-processing rewrite rule in single-pass analyzer. Because `ExprUtils.assertValidAggregate` is called during the bottom-up traversal, we can't check for non-determinstic expressions there

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51391 from mihailotim-db/mihailotim-db/pull_out_nondeterministic.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>